third_party/blink/web_tests/virtual/bidi-caret-affinity/editing/selection/caret-at-bidi-boundary-expected.txt

This test ensures WebKit lets user place caret around bidirectional text properly.

Current offset:
abcאבג
0 1 2 3 6 5 4 3
נת12
2 3 4 2 1 0
לשנת
4 3 2 1 0
aקל12יםd
0 1 7 6 5 3 4 5 3 2 1 7 8
12קל43ab
4 5 6 4 3 2 0 1 2 6 7 8
abcלשנ
0 1 2 3 6 5 4 3
Test "abcאבג":
PASS caret is at 0
PASS caret is at 1
PASS caret is at 2
PASS caret is at 3
PASS caret is at 6
PASS caret is at 5
PASS caret is at 4
PASS caret is at 3


Test "נת12":
PASS caret is at 2
PASS caret is at 3
PASS caret is at 4
PASS caret is at 2
PASS caret is at 1
PASS caret is at 0


Test "לשנת":
PASS caret is at 4
PASS caret is at 3
PASS caret is at 2
PASS caret is at 1
PASS caret is at 0


Test "aקל12יםd":
PASS caret is at 0
PASS caret is at 1
PASS caret is at 7
PASS caret is at 6
PASS caret is at 5
PASS caret is at 3
PASS caret is at 4
PASS caret is at 5
PASS caret is at 3
PASS caret is at 2
PASS caret is at 1
PASS caret is at 7
PASS caret is at 8


Test "12קל43ab":
PASS caret is at 4
PASS caret is at 5
PASS caret is at 6
PASS caret is at 4
PASS caret is at 3
PASS caret is at 2
PASS caret is at 0
PASS caret is at 1
PASS caret is at 2
PASS caret is at 6
PASS caret is at 7
PASS caret is at 8


Test "abcלשנ":
PASS caret is at 0
PASS caret is at 1
PASS caret is at 2
PASS caret is at 3
PASS caret is at 6
PASS caret is at 5
PASS caret is at 4
PASS caret is at 3


PASS successfullyParsed is true

TEST COMPLETE
